requirements.txt
streamlit
pymongo
openai
python-dotenv
nest_asyncio
